requirements.txt
Django==5.0.6
django-rest-framework==0.1.0
django-filter==24.2
djangorestframework==3.15.1
django-extensions==3.2.3
notebook==6.5.6